+++ changelog.md
@@ -5,8 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.3.0] - 2024-12-21
 
+- Print mnemonics sorted and in columns of 8
+- Add options --count and --begin-with
+- Remove option --repeat
 - Add logo
 
 ## [0.2.0] - 2022-12-04